Added main sample, washington_test generator and fixed bugs

+++ modules/data_structures/source/csr.cpp
@@ -9,8 +9,8 @@
 
 CSR::CSR(const AL& al, ...) {
     offset.push_back(0);
-    for (int i = 0; i < static_cast<int>(al.edges.size()); ++i) {
-        for (int j = 0; j < static_cast<int>(al.edges[i].size()); ++j) {
+    for (unsigned i = 0; i < al.edges.size(); ++i) {
+        for (unsigned j = 0; j < al.edges[i].size(); ++j) {
             edges.emplace_back(al.edges[i][j]);
         }
         offset.push_back(edges.size());
